refactor: normalize sort_order columns to INTEGER (B-011)
Change all sort_order columns from REAL to INTEGER since the code only
ever assigns whole numbers. Migration 016 handles the schema change for
pages, databases, database_records, and database_fields using SQLite's
DROP COLUMN + RENAME COLUMN (available since 3.25.0).

Also update in-memory test schema definitions to match.

+++ packages/server/src/handlers/pages.test.ts
@@ -22,7 +22,7 @@ const setupDB = Effect.gen(function* () {
       parent_id TEXT REFERENCES pages(id) ON DELETE SET NULL,
       icon TEXT,
       cover_url TEXT,
-      sort_order REAL NOT NULL DEFAULT 0,
+      sort_order INTEGER NOT NULL DEFAULT 0,
       is_deleted INTEGER NOT NULL DEFAULT 0,
       is_favorite INTEGER NOT NULL DEFAULT 0,
       created_at TEXT NOT NULL DEFAULT (datetime('now')),
